Codebase Cleanup and Optimisation
- Unified (no "official" suffixes)
- Descriptive headers in all files
- Enhanced retry logic with 15 attempts
- Updated cursorrules and AGENT.MD
- Two distinct logic paths for database operations
