+++ csv2notion_neo/__init__.py
@@ -1,0 +1,9 @@
+"""
+CSV2Notion Neo - Advanced CSV/JSON to Notion Database Upload Tool
+
+This module provides the main package initialization for CSV2Notion Neo,
+a sophisticated command-line tool for uploading and merging CSV or JSON files
+with images to Notion databases using the official Notion API.
+"""
+
+from csv2notion_neo.version import __version__
