+++ csv2notion_neo/__main__.py
@@ -1,3 +1,10 @@
+"""
+CSV2Notion Neo - Main Entry Point
+
+This module serves as the main entry point for the CSV2Notion Neo application.
+It provides a simple interface to launch the CLI application.
+"""
+
 from csv2notion_neo.cli import main
 from icecream import ic
