+++ csv2notion_neo/utils_threading.py
@@ -1,16 +1,24 @@
+"""
+CSV2Notion Neo - Threading Utilities
+
+This module provides threading utilities for concurrent operations in CSV2Notion Neo.
+It manages thread pools, concurrent uploads, and thread-safe operations to optimize
+performance when processing large datasets.
+"""
+
 import threading
 from concurrent.futures import ThreadPoolExecutor, as_completed
 from typing import Any, Callable, Iterable, Iterator
 
-from csv2notion_neo.notion_db_official import NotionDBOfficial
-from csv2notion_neo.notion_client_official import NotionClientOfficial
-from csv2notion_neo.notion_db_client_official import NotionClientExtendedOfficial
+from csv2notion_neo.notion_db import NotionDB
+from csv2notion_neo.notion_client import NotionClient
+from csv2notion_neo.notion_db_client import NotionClientExtended
 from csv2notion_neo.notion_uploader import NotionRowUploader
 from icecream import ic
 
 
 class ThreadRowUploader(object):
-    def __init__(self, client: NotionClientOfficial, collection_id: str) -> None:
+    def __init__(self, client: NotionClient, collection_id: str) -> None:
         self.thread_data = threading.local()
 
         self.client = client
@@ -20,8 +28,13 @@ class ThreadRowUploader(object):
         try:
             notion_uploader = self.thread_data.uploader
         except AttributeError:
-            client = NotionClientExtendedOfficial(old_client=self.client)
-            notion_db = NotionDBOfficial(client, self.collection_id)
+            # Create a new extended client for this thread
+            client = NotionClientExtended(
+                integration_token=self.client.integration_token,
+                workspace=self.client.workspace,
+                options=self.client.options
+            )
+            notion_db = NotionDB(client, self.collection_id)
 
             notion_uploader = NotionRowUploader(notion_db)
             self.thread_data.uploader = notion_uploader
